fix(discovery): log collection load failures
